Update static pages images to hanlde shared grafana data

+++ images/static/entrypoint.sh
@@ -1,4 +1,7 @@
 #!/bin/bash
-ls /grafana && rm -rf /usr/share/nginx/html/backups/grafana && mv /grafana /usr/share/nginx/html/backups/grafana
-mv replacer sqlitedb runq /usr/share/nginx/html/backups/grafana
+if [ ! -z "${REINIT_SHARED_GRAFANA}" ]
+then
+  ls /grafana && rm -rf /usr/share/nginx/html/backups/grafana && mv /grafana /usr/share/nginx/html/backups/grafana
+  mv replacer sqlitedb runq /usr/share/nginx/html/backups/grafana
+fi
 nginx -g 'daemon off;'
